Print elements of list

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -1,7 +1,6 @@
 #include "lists.h"
 #include <stdio.h>
 #include <string.h>
-#include <stdlib.h>
 /**
  * print_list - Function to print the list of a node
  * @h: The pointer to the nodes
@@ -9,7 +8,7 @@
  */
 size_t print_list(const list_t *h)
 {
-	int nodes = 0;
+	size_t nodes = 0;
 
 	while (h)
 	{
